Added proportional division of interest stream

+++ contracts/sai2dai/ScdMcdMigration.sol
@@ -1,10 +1,9 @@
 pragma solidity 0.5.17;
 
+
 interface ScdMcdMigration {
-  // Function to swap SAI to DAI
-  // This function is to be used by users that want to get new DAI in exchange of old one (aka SAI)
-  // wad amount has to be <= the value pending to reach the debt ceiling (the minimum between general and ilk one)
-  function swapSaiToDai(
-    uint wad
-  ) external;
+    // Function to swap SAI to DAI
+    // This function is to be used by users that want to get new DAI in exchange of old one (aka SAI)
+    // wad amount has to be <= the value pending to reach the debt ceiling (the minimum between general and ilk one)
+    function swapSaiToDai(uint256 wad) external;
 }
